requirements.txt
# python==3.9.6
pandas==2.2.2
numpy==1.26.4
tqdm==4.66.4
jupyter==1.0.0
typer==0.12.3
xgboost==2.1.0
scikit-learn==1.5.0
nip-config==0.10.3

# attention and cba_*
torch==2.6.0
# cba_*
huggingface_hub==0.28.1
transformers==4.48.3
# clustering
networkx==3.2.1
python-louvain==0.16